requirements.txt
discord.py>=2.3.2
requests>=2.31.0
aiosqlite>=0.19.0 